Added instructions to README.md to setup cloud run, including creating a secret key to authenticate requests. Added scripts/deploy_cloud_run.sh to deploy to cloud run. Modified scripts/deploy_gae.sh to use .app_engine_ignore. Modified scripts/deploy_cloud_function.sh to use .cloud_function_ignore. Moved validation functions from app_engine.py to validate.py Moved strip and sanitize functions from app_engine.py to strip.py Pass cloud_run_url and encoded_jwt to labelVideo.html/labelVideo.js and monitorTraining.html/monitorTraining.js. Added main_cloud_run.py with handlers for /retrieveVideoFrameEntitiesWithImageUrls and /retrieveSummaryItems. Use cloud_run_url for /retrieveVideoFrameEntitiesWithImageUrls and /retrieveSummaryItems requests.
