+++ scripts/deploy_cloud_function.sh
@@ -7,9 +7,10 @@ fi
 pushd server
 gcloud functions deploy perform_action \
     --runtime=python39 \
-    --set-env-vars PROJECT_ID=${FMLTC_GCLOUD_PROJECT_ID} \
+    --set-env-vars=PROJECT_ID=${FMLTC_GCLOUD_PROJECT_ID} \
     --memory=8192MB \
     --timeout=540 \
     --trigger-resource=${FMLTC_GCLOUD_PROJECT_ID}-action-parameters \
-    --trigger-event=google.storage.object.finalize
+    --trigger-event=google.storage.object.finalize \
+    --ignore-file=.cloud_function_ignore
 popd
