+++ server/requirements.txt
@@ -1,5 +1,7 @@
-Flask==2.0.2
 blinker==1.4
+Flask==2.0.2
+flask-cors==3.0.10
+flask-oidc-ext==1.4.3
 google-api-python-client==1.9.0
 google-cloud-datastore==2.1.0
 google-cloud-error-reporting
@@ -7,13 +9,13 @@ google-cloud-resource-manager==1.1.2
 google-cloud-secret-manager==2.7.0
 google-cloud-storage==1.35.0
 gunicorn==20.1.0
-flask-oidc-ext==1.4.3
 numpy==1.19.4
 object_detection-0.1_2.5.0.tar.gz
 opencv-contrib-python-headless==4.5.2.54
 Pillow==8.3.2
 protobuf==3.17.3
 psutil==5.8.0
+pyjwt==2.3.0
 python-dateutil==2.8.1
 redis==3.5.3
 sentry-sdk==1.4.3
